refactor(server): Split connection.rs into focused modules under 500 lines
Split the 1,775-line connection.rs into 8 focused modules:
- mod.rs (434 lines): ConnectionHandler struct and main connection logic
- auth.rs (151 lines): Authentication methods (trust, password, MD5)
- io.rs (29 lines): Low-level I/O operations
- protocol.rs (320 lines): Protocol message encoding/sending
- query.rs (128 lines): Query execution and result handling
- selective.rs (242 lines): Selective column update logic
- subscription.rs (463 lines): Subscription registration and updates
- updates.rs (300 lines): Subscription update sending logic

All modules stay under 500 lines (project guideline).
No functional changes - pure refactor.
All existing tests pass.
Public API unchanged.

Closes #4338
